Extract employee addition into a modal
